+++ Tests/ConnectTests/ActionContextTests.swift
@@ -34,7 +34,7 @@ class ActionContextTests: XCTestCase {
         super.setUp()
 
         do {
-            try removePersistentStoreCache()
+            try removePersistentStoreCache(for: modelName)
         } catch {
             XCTFail(error.localizedDescription)
         }