+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All significant changes to this project will be documented in this file.
 #### Added 
 - Added `ContextStrategyType` allowing different ManagedObjectContext layout strategies to be used for both `Connect` and `GenericPersistentContainer`.
 - Added `BackgroundContext` which is the base type returned from `ContextStrategy` `newBackgroundContext`.
+- Added `Configuration` class to allow configuration of Connect instances.
 
 #### Breaking Updates 
 - Changed `Connect` to generic class to support `ContextStrategyType`.